% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchData.R
\name{fetchData}
\alias{fetchData}
\title{fetchData}
\usage{
fetchData()
}
\value{
A tidy table containing daily information about coronavirus for all countries over the world
}
\description{
This function downloads updated coronavirus data. It contains time series of
confirmed and death cases of all countries in a tidy format.
It also contains geographic coordinates(Lat,Long) for each Region
}
\examples{
\dontrun{timeSeries <- fetchData()}

}
