% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountryTimeline.R
\name{plotCountryTimeline}
\alias{plotCountryTimeline}
\title{plotCountryTimeline}
\usage{
plotCountryTimeline(dateRange, country)
}
\arguments{
\item{dateRange}{A date vector of length 2. indicates a time period between two days}

\item{country}{Name of the country to be plotted. Tt should start with capital letter.}
}
\description{
plotCountryTimeline
}
\examples{
plotCountryTimeline(c("2020/2/05", "2020/08/27"), "US")

plotCountryTimeline(c("2020-10-01", "2020-10-29"), "Germany")

}
