% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWorldMap.R
\name{plotWorldMap}
\alias{plotWorldMap}
\title{plotWorldMap}
\usage{
plotWorldMap(dateRange, type)
}
\arguments{
\item{dateRange}{a date vector of length 1 or 2. if one element is passed, indicates a
single day. if two elements are passed, indicates a sequence of days between two days.}

\item{type}{type of the coronavirus case. one of "Confirmed" or "Death"}
}
\description{
This function plots a choropleth map of the world showing the spread
of coronavirus in a specific time period
}
\examples{
\dontrun{plotWorldMap(c("2020/03/26","2020/05/12"), type = "Confirmed")}
\dontrun{plotWorldMap(c("2020/06/21"), type = "Death")}

}
